# src/modules/allocations/allocations.controller.ts
import { inject } from 'inversify';
// eslint has a trouble to see that 'Query' is actually used ¯\_(ツ)_/¯
// eslint-disable-next-line @typescript-eslint/no-unused-vars
import { Body, Delete, Get, Post, Put, Query, Route, Security } from 'tsoa';
import { provideSingleton } from '~/ioc/container';
import { assertIsValid, UserError } from '~/modules/common';

import { AllocationService } from './allocations.service';
import { AllocationCreateRequest, AllocationUpdateRequest, AllocationResponse } from './allocations.dto';

@Route('allocations')
@provideSingleton(AllocationsController)
export class AllocationsController {
  constructor(@inject(AllocationService) private readonly service: AllocationService) {}

  /**
   *
   * @isInt skip
   * @isInt take
   * @minimum  take 0
   * @minimum  skip 0
   * @maximum take 100
   */
  @Get()
  @Security('jwt')
  public async getAll(@Query() skip: number = 0, @Query() take: number = 100): Promise<AllocationResponse[]> {
    return await this.service.readAll({ take, skip });
  }

  /**
   * @param uuid
   * @pattern uuid [0-9a-fA-F]{8}\-[0-9a-fA-F]{4}\-[0-9a-fA-F]{4}\-[0-9a-fA-F]{4}\-[0-9a-fA-F]{12}
   */
  @Get('{uuid}')
  @Security('jwt')
  public async get(uuid: string): Promise<AllocationResponse> {
    const result = await this.service.read(uuid);
    if (!result) {
      throw new UserError(404, `Allocation with uuid '${uuid}' was not found`);
    }
    return result;
  }

  @Post()
  @Security('jwt')
  public async create(@Body() request: AllocationCreateRequest): Promise<AllocationResponse> {
    await assertIsValid(Object.assign(new AllocationCreateRequest(), request));
    const result = await this.service.create(request);
    return result;
  }

  /**
   * @param uuid
   * @pattern uuid [0-9a-fA-F]{8}\-[0-9a-fA-F]{4}\-[0-9a-fA-F]{4}\-[0-9a-fA-F]{4}\-[0-9a-fA-F]{12}
   */
  @Put('{uuid}')
  @Security('jwt')
  public async update(uuid: string, @Body() request: AllocationUpdateRequest): Promise<AllocationResponse> {
    await assertIsValid(Object.assign(new AllocationUpdateRequest(), request));
    const result = await this.service.update(uuid, request);
    return result;
  }

  /**
   * @returns {number} Number of affected rows/documents
   * @param uuid
   * @pattern uuid [0-9a-fA-F]{8}\-[0-9a-fA-F]{4}\-[0-9a-fA-F]{4}\-[0-9a-fA-F]{4}\-[0-9a-fA-F]{12}
   */
  @Delete('{uuid}')
  @Security('jwt')
  public async delete(uuid: string): Promise<number> {
    const result = await this.service.delete(uuid);
    if (result.affected === 0) {
      throw new UserError(404, `Allocation with uuid '${uuid}' was not found`);
    }
    return result.affected;
  }
}