# src/server/swagger.json
{
	"basePath": "/",
	"consumes": [
		"application/json"
	],
	"definitions": {
		"UserResponse": {
			"properties": {
				"uuid": {
					"type": "string"
				},
				"email": {
					"type": "string"
				},
				"firstName": {
					"type": "string",
					"x-nullable": true
				},
				"lastName": {
					"type": "string",
					"x-nullable": true
				}
			},
			"required": [
				"uuid",
				"email"
			],
			"type": "object"
		},
		"UserCreateRequest": {
			"properties": {
				"email": {
					"type": "string"
				},
				"password": {
					"type": "string"
				},
				"firstName": {
					"type": "string",
					"x-nullable": true
				},
				"lastName": {
					"type": "string",
					"x-nullable": true
				}
			},
			"required": [
				"email",
				"password"
			],
			"type": "object"
		},
		"UserUpdateRequest": {
			"properties": {
				"email": {
					"type": "string",
					"x-nullable": true
				},
				"password": {
					"type": "string",
					"x-nullable": true
				},
				"firstName": {
					"type": "string",
					"x-nullable": true
				},
				"lastName": {
					"type": "string",
					"x-nullable": true
				}
			},
			"type": "object"
		},
		"AuthPayload": {
			"properties": {
				"email": {
					"type": "string"
				},
				"password": {
					"type": "string"
				}
			},
			"required": [
				"email",
				"password"
			],
			"type": "object"
		},
		"AssetTypeResponse": {
			"properties": {
				"uuid": {
					"type": "string"
				},
				"code": {
					"type": "string"
				},
				"attributes": {
					"type": "array",
					"items": {
						"type": "string"
					}
				}
			},
			"required": [
				"uuid",
				"code",
				"attributes"
			],
			"type": "object"
		},
		"AssetResponse": {
			"properties": {
				"name": {
					"type": "string"
				},
				"type": {
					"$ref": "#/definitions/AssetTypeResponse"
				},
				"attributeValues": {
					"type": "object"
				}
			},
			"required": [
				"name",
				"type",
				"attributeValues"
			],
			"type": "object"
		},
		"AssetCreateRequest": {
			"properties": {
				"typeCode": {
					"type": "string"
				},
				"name": {
					"type": "string"
				},
				"attributeValues": {
					"type": "object"
				}
			},
			"required": [
				"typeCode",
				"name",
				"attributeValues"
			],
			"type": "object"
		},
		"AssetUpdateRequest": {
			"properties": {
				"typeCode": {
					"type": "string",
					"x-nullable": true
				},
				"name": {
					"type": "string",
					"x-nullable": true
				},
				"attributeValues": {
					"type": "object",
					"x-nullable": true
				}
			},
			"type": "object"
		},
		"AssetTypeCreateRequest": {
			"properties": {
				"code": {
					"type": "string"
				},
				"attributes": {
					"type": "array",
					"items": {
						"type": "string"
					}
				}
			},
			"required": [
				"code",
				"attributes"
			],
			"type": "object"
		},
		"AssetTypeUpdateRequest": {
			"properties": {
				"code": {
					"type": "string"
				},
				"attributes": {
					"type": "array",
					"items": {
						"type": "string"
					}
				}
			},
			"required": [
				"code",
				"attributes"
			],
			"type": "object"
		},
		"AllocationResponse": {
			"properties": {},
			"type": "object"
		},
		"AllocationCreateRequest": {
			"properties": {},
			"type": "object"
		},
		"AllocationUpdateRequest": {
			"properties": {},
			"type": "object"
		}
	},
	"info": {
		"title": "bulb-server",
		"version": "0.1.0",
		"description": "Assets manager"
	},
	"paths": {
		"/users": {
			"get": {
				"operationId": "GetAll",
				"produces": [
					"application/json"
				],
				"responses": {
					"200": {
						"description": "Ok",
						"schema": {
							"type": "array",
							"items": {
								"$ref": "#/definitions/UserResponse"
							}
						}
					}
				},
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": [
					{
						"default": 0,
						"in": "query",
						"name": "skip",
						"required": false,
						"format": "int32",
						"type": "integer",
						"minimum": 0
					},
					{
						"default": 100,
						"in": "query",
						"name": "take",
						"required": false,
						"format": "int32",
						"type": "integer",
						"minimum": 0,
						"maximum": 100
					}
				]
			},
			"post": {
				"operationId": "Create",
				"produces": [
					"application/json"
				],
				"responses": {
					"200": {
						"description": "Ok",
						"schema": {
							"$ref": "#/definitions/UserResponse"
						}
					}
				},
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": [
					{
						"in": "body",
						"name": "request",
						"required": true,
						"schema": {
							"$ref": "#/definitions/UserCreateRequest"
						}
					}
				]
			}
		},
		"/users/{uuid}": {
			"get": {
				"operationId": "Get",
				"produces": [
					"application/json"
				],
				"responses": {
					"200": {
						"description": "Ok",
						"schema": {
							"$ref": "#/definitions/UserResponse"
						}
					}
				},
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "uuid",
						"required": true,
						"type": "string",
						"pattern": "[0-9a-fA-F]{8}\\-[0-9a-fA-F]{4}\\-[0-9a-fA-F]{4}\\-[0-9a-fA-F]{4}\\-[0-9a-fA-F]{12}"
					}
				]
			},
			"put": {
				"operationId": "Update",
				"produces": [
					"application/json"
				],
				"responses": {
					"200": {
						"description": "Ok",
						"schema": {
							"$ref": "#/definitions/UserResponse"
						}
					}
				},
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "uuid",
						"required": true,
						"type": "string",
						"pattern": "[0-9a-fA-F]{8}\\-[0-9a-fA-F]{4}\\-[0-9a-fA-F]{4}\\-[0-9a-fA-F]{4}\\-[0-9a-fA-F]{12}"
					},
					{
						"in": "body",
						"name": "request",
						"required": true,
						"schema": {
							"$ref": "#/definitions/UserUpdateRequest"
						}
					}
				]
			},
			"delete": {
				"operationId": "Delete",
				"produces": [
					"application/json"
				],
				"responses": {
					"200": {
						"description": "Ok",
						"schema": {
							"type": "number",
							"format": "double"
						}
					}
				},
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "uuid",
						"required": true,
						"type": "string",
						"pattern": "[0-9a-fA-F]{8}\\-[0-9a-fA-F]{4}\\-[0-9a-fA-F]{4}\\-[0-9a-fA-F]{4}\\-[0-9a-fA-F]{12}"
					}
				]
			}
		},
		"/auth/login": {
			"post": {
				"operationId": "Login",
				"produces": [
					"application/json"
				],
				"responses": {
					"200": {
						"description": "Ok",
						"schema": {
							"type": "object"
						}
					}
				},
				"security": [],
				"parameters": [
					{
						"in": "body",
						"name": "model",
						"required": true,
						"schema": {
							"$ref": "#/definitions/AuthPayload"
						}
					}
				]
			}
		},
		"/auth/signup": {
			"post": {
				"operationId": "Signup",
				"produces": [
					"application/json"
				],
				"responses": {
					"200": {
						"description": "Ok",
						"schema": {
							"$ref": "#/definitions/UserResponse"
						}
					}
				},
				"security": [],
				"parameters": [
					{
						"in": "body",
						"name": "request",
						"required": true,
						"schema": {
							"$ref": "#/definitions/UserCreateRequest"
						}
					}
				]
			}
		},
		"/assets": {
			"get": {
				"operationId": "GetAll",
				"produces": [
					"application/json"
				],
				"responses": {
					"200": {
						"description": "Ok",
						"schema": {
							"type": "array",
							"items": {
								"$ref": "#/definitions/AssetResponse"
							}
						}
					}
				},
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": [
					{
						"default": 0,
						"in": "query",
						"name": "skip",
						"required": false,
						"format": "int32",
						"type": "integer",
						"minimum": 0
					},
					{
						"default": 100,
						"in": "query",
						"name": "take",
						"required": false,
						"format": "int32",
						"type": "integer",
						"minimum": 0,
						"maximum": 100
					}
				]
			},
			"post": {
				"operationId": "Create",
				"produces": [
					"application/json"
				],
				"responses": {
					"200": {
						"description": "Ok",
						"schema": {
							"$ref": "#/definitions/AssetResponse"
						}
					}
				},
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": [
					{
						"in": "body",
						"name": "request",
						"required": true,
						"schema": {
							"$ref": "#/definitions/AssetCreateRequest"
						}
					}
				]
			}
		},
		"/assets/{uuid}": {
			"get": {
				"operationId": "Get",
				"produces": [
					"application/json"
				],
				"responses": {
					"200": {
						"description": "Ok",
						"schema": {
							"$ref": "#/definitions/AssetResponse"
						}
					}
				},
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "uuid",
						"required": true,
						"type": "string",
						"pattern": "[0-9a-fA-F]{8}\\-[0-9a-fA-F]{4}\\-[0-9a-fA-F]{4}\\-[0-9a-fA-F]{4}\\-[0-9a-fA-F]{12}"
					}
				]
			},
			"put": {
				"operationId": "Update",
				"produces": [
					"application/json"
				],
				"responses": {
					"200": {
						"description": "Ok",
						"schema": {
							"$ref": "#/definitions/AssetResponse"
						}
					}
				},
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "uuid",
						"required": true,
						"type": "string",
						"pattern": "[0-9a-fA-F]{8}\\-[0-9a-fA-F]{4}\\-[0-9a-fA-F]{4}\\-[0-9a-fA-F]{4}\\-[0-9a-fA-F]{12}"
					},
					{
						"in": "body",
						"name": "request",
						"required": true,
						"schema": {
							"$ref": "#/definitions/AssetUpdateRequest"
						}
					}
				]
			},
			"delete": {
				"operationId": "Delete",
				"produces": [
					"application/json"
				],
				"responses": {
					"200": {
						"description": "Ok",
						"schema": {
							"type": "number",
							"format": "double"
						}
					}
				},
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "uuid",
						"required": true,
						"type": "string",
						"pattern": "[0-9a-fA-F]{8}\\-[0-9a-fA-F]{4}\\-[0-9a-fA-F]{4}\\-[0-9a-fA-F]{4}\\-[0-9a-fA-F]{12}"
					}
				]
			}
		},
		"/assetTypes": {
			"get": {
				"operationId": "GetAll",
				"produces": [
					"application/json"
				],
				"responses": {
					"200": {
						"description": "Ok",
						"schema": {
							"type": "array",
							"items": {
								"$ref": "#/definitions/AssetTypeResponse"
							}
						}
					}
				},
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": [
					{
						"default": 0,
						"in": "query",
						"name": "skip",
						"required": false,
						"format": "int32",
						"type": "integer",
						"minimum": 0
					},
					{
						"default": 100,
						"in": "query",
						"name": "take",
						"required": false,
						"format": "int32",
						"type": "integer",
						"minimum": 0,
						"maximum": 100
					}
				]
			},
			"post": {
				"operationId": "Create",
				"produces": [
					"application/json"
				],
				"responses": {
					"200": {
						"description": "Ok",
						"schema": {
							"$ref": "#/definitions/AssetTypeResponse"
						}
					}
				},
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": [
					{
						"in": "body",
						"name": "request",
						"required": true,
						"schema": {
							"$ref": "#/definitions/AssetTypeCreateRequest"
						}
					}
				]
			}
		},
		"/assetTypes/{uuidOrCode}": {
			"get": {
				"operationId": "Get",
				"produces": [
					"application/json"
				],
				"responses": {
					"200": {
						"description": "Ok",
						"schema": {
							"$ref": "#/definitions/AssetTypeResponse"
						}
					}
				},
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "uuidOrCode",
						"required": true,
						"type": "string",
						"minLength": 2
					}
				]
			}
		},
		"/assetTypes/{uuid}": {
			"put": {
				"operationId": "Update",
				"produces": [
					"application/json"
				],
				"responses": {
					"200": {
						"description": "Ok",
						"schema": {
							"$ref": "#/definitions/AssetTypeResponse"
						}
					}
				},
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "uuid",
						"required": true,
						"type": "string",
						"pattern": "[0-9a-fA-F]{8}\\-[0-9a-fA-F]{4}\\-[0-9a-fA-F]{4}\\-[0-9a-fA-F]{4}\\-[0-9a-fA-F]{12}"
					},
					{
						"in": "body",
						"name": "request",
						"required": true,
						"schema": {
							"$ref": "#/definitions/AssetTypeUpdateRequest"
						}
					}
				]
			},
			"delete": {
				"operationId": "Delete",
				"produces": [
					"application/json"
				],
				"responses": {
					"200": {
						"description": "Ok",
						"schema": {
							"type": "number",
							"format": "double"
						}
					}
				},
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "uuid",
						"required": true,
						"type": "string",
						"pattern": "[0-9a-fA-F]{8}\\-[0-9a-fA-F]{4}\\-[0-9a-fA-F]{4}\\-[0-9a-fA-F]{4}\\-[0-9a-fA-F]{12}"
					}
				]
			}
		},
		"/allocations": {
			"get": {
				"operationId": "GetAll",
				"produces": [
					"application/json"
				],
				"responses": {
					"200": {
						"description": "Ok",
						"schema": {
							"type": "array",
							"items": {
								"$ref": "#/definitions/AllocationResponse"
							}
						}
					}
				},
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": [
					{
						"default": 0,
						"in": "query",
						"name": "skip",
						"required": false,
						"format": "int32",
						"type": "integer",
						"minimum": 0
					},
					{
						"default": 100,
						"in": "query",
						"name": "take",
						"required": false,
						"format": "int32",
						"type": "integer",
						"minimum": 0,
						"maximum": 100
					}
				]
			},
			"post": {
				"operationId": "Create",
				"produces": [
					"application/json"
				],
				"responses": {
					"200": {
						"description": "Ok",
						"schema": {
							"$ref": "#/definitions/AllocationResponse"
						}
					}
				},
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": [
					{
						"in": "body",
						"name": "request",
						"required": true,
						"schema": {
							"$ref": "#/definitions/AllocationCreateRequest"
						}
					}
				]
			}
		},
		"/allocations/{uuid}": {
			"get": {
				"operationId": "Get",
				"produces": [
					"application/json"
				],
				"responses": {
					"200": {
						"description": "Ok",
						"schema": {
							"$ref": "#/definitions/AllocationResponse"
						}
					}
				},
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "uuid",
						"required": true,
						"type": "string",
						"pattern": "[0-9a-fA-F]{8}\\-[0-9a-fA-F]{4}\\-[0-9a-fA-F]{4}\\-[0-9a-fA-F]{4}\\-[0-9a-fA-F]{12}"
					}
				]
			},
			"put": {
				"operationId": "Update",
				"produces": [
					"application/json"
				],
				"responses": {
					"200": {
						"description": "Ok",
						"schema": {
							"$ref": "#/definitions/AllocationResponse"
						}
					}
				},
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "uuid",
						"required": true,
						"type": "string",
						"pattern": "[0-9a-fA-F]{8}\\-[0-9a-fA-F]{4}\\-[0-9a-fA-F]{4}\\-[0-9a-fA-F]{4}\\-[0-9a-fA-F]{12}"
					},
					{
						"in": "body",
						"name": "request",
						"required": true,
						"schema": {
							"$ref": "#/definitions/AllocationUpdateRequest"
						}
					}
				]
			},
			"delete": {
				"operationId": "Delete",
				"produces": [
					"application/json"
				],
				"responses": {
					"200": {
						"description": "Ok",
						"schema": {
							"type": "number",
							"format": "double"
						}
					}
				},
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "uuid",
						"required": true,
						"type": "string",
						"pattern": "[0-9a-fA-F]{8}\\-[0-9a-fA-F]{4}\\-[0-9a-fA-F]{4}\\-[0-9a-fA-F]{4}\\-[0-9a-fA-F]{12}"
					}
				]
			}
		}
	},
	"produces": [
		"application/json"
	],
	"swagger": "2.0",
	"securityDefinitions": {
		"jwt": {
			"type": "apiKey",
			"name": "authorization",
			"in": "header"
		}
	},
	"host": "localhost:3010"
}